Develop one to one with component (#2753)
* Made Policy Factory a component

* Broke the Heuristic

* fix bug

* BugFix

* Reimplemented Heuristic

* Changing all of the prefabs

* forgotten file

* Fixing under which conditions the Heuristic policy is used

* Dispose of brain
