+++ UnitySDK/Assets/ML-Agents/Editor/AgentEditor.cs
@@ -11,12 +11,6 @@ namespace MLAgents
     [CanEditMultipleObjects]
     public class AgentEditor : Editor
     {
-        private const float k_TimeBetweenModelReloads = 2f;
-        // Time since the last reload of the model
-        private float m_TimeSinceModelReload;
-        // Whether or not the model needs to be reloaded
-        private bool m_RequireReload;
-
         public override void OnInspectorGUI()
         {
             var serializedAgent = serializedObject;
@@ -34,29 +28,6 @@ namespace MLAgents
                 "agentParameters.agentCameras");
             var renderTextures = serializedAgent.FindProperty(
                 "agentParameters.agentRenderTextures");
-
-            // Drawing the Behavior Parameters
-            var brainParameters = serializedAgent.FindProperty("m_BrainParameters");
-            brainParameters.isExpanded = EditorGUILayout.Foldout(brainParameters.isExpanded, "Behavior Parameters");
-            if (brainParameters.isExpanded)
-            {
-                EditorGUI.BeginChangeCheck();
-                EditorGUI.indentLevel++;
-                EditorGUILayout.PropertyField(serializedAgent.FindProperty("m_BehaviorName"));
-                EditorGUILayout.PropertyField(serializedAgent.FindProperty("m_BrainParameters"), true);
-                EditorGUILayout.PropertyField(serializedAgent.FindProperty("m_Model"), true);
-                EditorGUI.indentLevel++;
-                EditorGUILayout.PropertyField(serializedAgent.FindProperty("m_InferenceDevice"), true);
-                EditorGUI.indentLevel--;
-                EditorGUILayout.PropertyField(serializedAgent.FindProperty("m_UseHeuristic"));
-                EditorGUI.indentLevel--;
-                if (EditorGUI.EndChangeCheck())
-                {
-                    m_RequireReload = true;
-                }
-                DisplayFailedModelChecks();
-            }
-            EditorGUILayout.LabelField("", GUI.skin.horizontalSlider);
 
 
             if (cameras.arraySize > 0 && renderTextures.arraySize > 0)
@@ -136,39 +107,6 @@ namespace MLAgents
 
             EditorGUILayout.LabelField("", GUI.skin.horizontalSlider);
             base.OnInspectorGUI();
-        }
-
-        /// <summary>
-        /// Must be called within OnEditorGUI()
-        /// </summary>
-        private void DisplayFailedModelChecks()
-        {
-            if (m_RequireReload && m_TimeSinceModelReload > k_TimeBetweenModelReloads)
-            {
-                m_RequireReload = false;
-                m_TimeSinceModelReload = 0;
-            }
-            // Display all failed checks
-            D.logEnabled = false;
-            Model barracudaModel = null;
-            var model = (NNModel)serializedObject.FindProperty("m_Model").objectReferenceValue;
-            var brainParameters = ((Agent)target).brainParameters;
-            if (model != null)
-            {
-                barracudaModel = ModelLoader.Load(model.Value);
-            }
-            if (brainParameters != null)
-            {
-                var failedChecks = InferenceBrain.BarracudaModelParamLoader.CheckModel(
-                    barracudaModel, brainParameters);
-                foreach (var check in failedChecks)
-                {
-                    if (check != null)
-                    {
-                        EditorGUILayout.HelpBox(check, MessageType.Warning);
-                    }
-                }
-            }
         }
     }
 }
