+++ UnitySDK/Assets/ML-Agents/Editor/BehaviorParametersEditor.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 72b0b21a2d4ee4bc2be0530fd134720d
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
