+++ UnitySDK/Assets/ML-Agents/Examples/3DBall/Scenes/3DBall.unity
@@ -424,31 +424,6 @@ Prefab:
   m_Modification:
     m_TransformParent: {fileID: 0}
     m_Modifications:
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: agentParameters.agentCameras.Array.size
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.keyContinuousPlayerActions.Array.size
-      value: 3
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.vectorActionSize.Array.size
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.size
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.vectorActionDescriptions.Array.size
-      value: 1
-      objectReference: {fileID: 0}
     - target: {fileID: 4679453577574622, guid: cfa81c019162c4e3caf6e2999c6fdf48, type: 2}
       propertyPath: m_LocalPosition.x
       value: 0
@@ -479,110 +454,7 @@ Prefab:
       objectReference: {fileID: 0}
     - target: {fileID: 4679453577574622, guid: cfa81c019162c4e3caf6e2999c6fdf48, type: 2}
       propertyPath: m_RootOrder
-      value: 4
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.data[0].width
-      value: 84
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.data[0].height
-      value: 84
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.data[1].width
-      value: 84
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.data[1].height
-      value: 84
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brain
-      value: 
-      objectReference: {fileID: 11400000, guid: 383c589e8bb76464eadc2525b5b0f2c1,
-        type: 2}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.model
-      value: 
-      objectReference: {fileID: 11400000, guid: c60a63ad5dc0c4a029d7360054667457,
-        type: 3}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.inferenceDevice
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.data[2].width
-      value: 84
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.data[2].height
-      value: 84
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.data[3].width
-      value: 84
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.data[3].height
-      value: 84
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.data[4].width
-      value: 84
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.data[4].height
-      value: 84
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.vectorActionSize.Array.data[1]
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.vectorActionSize.Array.data[2]
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.vectorActionSpaceType
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.data[1].blackAndWhite
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.brainParameters.cameraResolutions.Array.data[0].blackAndWhite
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: brainFactoryParameters.decisionScript
-      value: 
-      objectReference: {fileID: 11500000, guid: ccaa8f43c15264209b137d8dc26a8d63,
-        type: 3}
-    - target: {fileID: 114259948429386406, guid: cfa81c019162c4e3caf6e2999c6fdf48,
-        type: 2}
-      propertyPath: m_BrainFactoryParameters.useHeuristic
-      value: 0
+      value: 5
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: cfa81c019162c4e3caf6e2999c6fdf48, type: 2}
@@ -869,6 +741,10 @@ Prefab:
       propertyPath: m_Name
       value: 3DBall (1)
       objectReference: {fileID: 0}
+    - target: {fileID: 0}
+      propertyPath: m_BrainParameters.numStackedVectorObservations
+      value: 1
+      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_ParentPrefab: {fileID: 100100000, guid: cfa81c019162c4e3caf6e2999c6fdf48, type: 2}
   m_IsPrefabParent: 0
@@ -1001,7 +877,7 @@ Transform:
   m_LocalScale: {x: 5, y: 0.19999993, z: 5}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 5
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!114 &1583402089
 MonoBehaviour:
