+++ UnitySDK/Assets/ML-Agents/Examples/PushBlock/Prefabs/PushBlockArea.prefab
@@ -104,6 +104,7 @@ GameObject:
   m_Component:
   - component: {fileID: 4188187884171146}
   - component: {fileID: 54817351390947638}
+  - component: {fileID: 114306175693660464}
   - component: {fileID: 114505490781873732}
   - component: {fileID: 114421647563711602}
   - component: {fileID: 65880096262939968}
@@ -933,6 +934,28 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   agent: {fileID: 0}
+--- !u!114 &114306175693660464
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1489716781518988}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5d1c4e0b1822b495aa52bc52839ecb30, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_BrainParameters:
+    vectorObservationSize: 70
+    numStackedVectorObservations: 3
+    vectorActionSize: 07000000
+    cameraResolutions: []
+    vectorActionDescriptions: []
+    vectorActionSpaceType: 0
+  m_Model: {fileID: 11400000, guid: c60a63ad5dc0c4a029d7360054667457, type: 3}
+  m_InferenceDevice: 0
+  m_UseHeuristic: 0
+  m_BehaviorName: PushBlock
 --- !u!114 &114421647563711602
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -955,17 +978,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: dea8c4f2604b947e6b7b97750dde87ca, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_BrainParameters:
-    vectorObservationSize: 70
-    numStackedVectorObservations: 3
-    vectorActionSize: 07000000
-    cameraResolutions: []
-    vectorActionDescriptions: []
-    vectorActionSpaceType: 0
-  m_Model: {fileID: 11400000, guid: c60a63ad5dc0c4a029d7360054667457, type: 3}
-  m_InferenceDevice: 0
-  m_UseHeuristic: 0
-  m_BehaviorName: PushBlock
   agentParameters:
     agentCameras: []
     agentRenderTextures: []
