+++ UnitySDK/Assets/ML-Agents/Examples/Pyramids/Prefabs/AreaPB.prefab
@@ -128,6 +128,7 @@ GameObject:
   m_Component:
   - component: {fileID: 4518417139497368}
   - component: {fileID: 54596704247224538}
+  - component: {fileID: 114399412043818042}
   - component: {fileID: 114937736047215868}
   - component: {fileID: 114507422577425370}
   - component: {fileID: 65345930959735878}
@@ -2872,6 +2873,28 @@ MonoBehaviour:
   - {fileID: 1589816231338102}
   numPyra: 1
   range: 45
+--- !u!114 &114399412043818042
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1131043459059966}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5d1c4e0b1822b495aa52bc52839ecb30, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_BrainParameters:
+    vectorObservationSize: 172
+    numStackedVectorObservations: 1
+    vectorActionSize: 05000000
+    cameraResolutions: []
+    vectorActionDescriptions: []
+    vectorActionSpaceType: 0
+  m_Model: {fileID: 11400000, guid: 9bafa731bfcbc4f0faa73c365e7af924, type: 3}
+  m_InferenceDevice: 0
+  m_UseHeuristic: 0
+  m_BehaviorName: Pyramids
 --- !u!114 &114507422577425370
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -2908,17 +2931,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b8db44472779248d3be46895c4d562d5, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_BrainParameters:
-    vectorObservationSize: 172
-    numStackedVectorObservations: 1
-    vectorActionSize: 05000000
-    cameraResolutions: []
-    vectorActionDescriptions: []
-    vectorActionSpaceType: 0
-  m_Model: {fileID: 11400000, guid: 9bafa731bfcbc4f0faa73c365e7af924, type: 3}
-  m_InferenceDevice: 0
-  m_UseHeuristic: 0
-  m_BehaviorName: Pyramids
   agentParameters:
     agentCameras: []
     agentRenderTextures: []
