+++ UnitySDK/Assets/ML-Agents/Examples/Soccer/Prefabs/SoccerFieldTwos.prefab
@@ -87,6 +87,7 @@ GameObject:
   - component: {fileID: 4277721046484044}
   - component: {fileID: 54348679551516588}
   - component: {fileID: 135232974003521068}
+  - component: {fileID: 114734187185382186}
   - component: {fileID: 114492261207303438}
   - component: {fileID: 114692966630797794}
   m_Layer: 13
@@ -177,6 +178,7 @@ GameObject:
   - component: {fileID: 4485793831109164}
   - component: {fileID: 54250052574815742}
   - component: {fileID: 135154818167532598}
+  - component: {fileID: 114105115387635628}
   - component: {fileID: 114698199869072806}
   - component: {fileID: 114381244552195858}
   m_Layer: 11
@@ -196,6 +198,7 @@ GameObject:
   - component: {fileID: 4444285537983296}
   - component: {fileID: 54609996481602788}
   - component: {fileID: 135208952479003512}
+  - component: {fileID: 114387866097048300}
   - component: {fileID: 114850431417842684}
   - component: {fileID: 114965771318032104}
   m_Layer: 13
@@ -761,6 +764,7 @@ GameObject:
   - component: {fileID: 4002186104597906}
   - component: {fileID: 54629836435839708}
   - component: {fileID: 135133947297127334}
+  - component: {fileID: 114529615399004778}
   - component: {fileID: 114284769194328828}
   - component: {fileID: 114724674330921748}
   m_Layer: 11
@@ -3505,6 +3509,28 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 30, y: 0.1, z: 16}
   m_Center: {x: 0, y: 0, z: 0}
+--- !u!114 &114105115387635628
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1124213441168130}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5d1c4e0b1822b495aa52bc52839ecb30, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_BrainParameters:
+    vectorObservationSize: 112
+    numStackedVectorObservations: 1
+    vectorActionSize: 05000000
+    cameraResolutions: []
+    vectorActionDescriptions: []
+    vectorActionSpaceType: 0
+  m_Model: {fileID: 0}
+  m_InferenceDevice: 0
+  m_UseHeuristic: 0
+  m_BehaviorName: Goalie
 --- !u!114 &114273807544954564
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -3532,17 +3558,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2a2688ef4a36349f9aa010020c32d198, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_BrainParameters:
-    vectorObservationSize: 112
-    numStackedVectorObservations: 1
-    vectorActionSize: 05000000
-    cameraResolutions: []
-    vectorActionDescriptions: []
-    vectorActionSpaceType: 0
-  m_Model: {fileID: 0}
-  m_InferenceDevice: 0
-  m_UseHeuristic: 0
-  m_BehaviorName: Goalie
   agentParameters:
     agentCameras: []
     agentRenderTextures: []
@@ -3565,15 +3580,15 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bb172294dbbcc408286b156a2c4b553c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &114492261207303438
+--- !u!114 &114387866097048300
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1095606497496374}
+  m_GameObject: {fileID: 1131626411948014}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2a2688ef4a36349f9aa010020c32d198, type: 3}
+  m_Script: {fileID: 11500000, guid: 5d1c4e0b1822b495aa52bc52839ecb30, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   m_BrainParameters:
@@ -3587,6 +3602,17 @@ MonoBehaviour:
   m_InferenceDevice: 0
   m_UseHeuristic: 0
   m_BehaviorName: Striker
+--- !u!114 &114492261207303438
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1095606497496374}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2a2688ef4a36349f9aa010020c32d198, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   agentParameters:
     agentCameras: []
     agentRenderTextures: []
@@ -3598,6 +3624,28 @@ MonoBehaviour:
   agentRole: 0
   area: {fileID: 114559182131992928}
   agentRb: {fileID: 0}
+--- !u!114 &114529615399004778
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1890219402901316}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5d1c4e0b1822b495aa52bc52839ecb30, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_BrainParameters:
+    vectorObservationSize: 112
+    numStackedVectorObservations: 1
+    vectorActionSize: 05000000
+    cameraResolutions: []
+    vectorActionDescriptions: []
+    vectorActionSpaceType: 0
+  m_Model: {fileID: 0}
+  m_InferenceDevice: 0
+  m_UseHeuristic: 0
+  m_BehaviorName: Goalie
 --- !u!114 &114559182131992928
 MonoBehaviour:
   m_ObjectHideFlags: 1
@@ -3639,17 +3687,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2a2688ef4a36349f9aa010020c32d198, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_BrainParameters:
-    vectorObservationSize: 112
-    numStackedVectorObservations: 1
-    vectorActionSize: 05000000
-    cameraResolutions: []
-    vectorActionDescriptions: []
-    vectorActionSpaceType: 0
-  m_Model: {fileID: 0}
-  m_InferenceDevice: 0
-  m_UseHeuristic: 0
-  m_BehaviorName: Goalie
   agentParameters:
     agentCameras: []
     agentRenderTextures: []
@@ -3672,15 +3709,15 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: bb172294dbbcc408286b156a2c4b553c, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
---- !u!114 &114850431417842684
+--- !u!114 &114734187185382186
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1131626411948014}
+  m_GameObject: {fileID: 1095606497496374}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2a2688ef4a36349f9aa010020c32d198, type: 3}
+  m_Script: {fileID: 11500000, guid: 5d1c4e0b1822b495aa52bc52839ecb30, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   m_BrainParameters:
@@ -3694,6 +3731,17 @@ MonoBehaviour:
   m_InferenceDevice: 0
   m_UseHeuristic: 0
   m_BehaviorName: Striker
+--- !u!114 &114850431417842684
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1131626411948014}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2a2688ef4a36349f9aa010020c32d198, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   agentParameters:
     agentCameras: []
     agentRenderTextures: []
