+++ UnitySDK/Assets/ML-Agents/Examples/Tennis/Prefabs/TennisArea.prefab
@@ -40,6 +40,7 @@ GameObject:
   - component: {fileID: 23050935508163814}
   - component: {fileID: 54815576193067388}
   - component: {fileID: 65276341973995358}
+  - component: {fileID: 114176423636690854}
   - component: {fileID: 114915946461826994}
   m_Layer: 0
   m_Name: AgentA
@@ -297,6 +298,7 @@ GameObject:
   - component: {fileID: 23268445935516234}
   - component: {fileID: 54459681652844648}
   - component: {fileID: 65280384434867516}
+  - component: {fileID: 114399072728845634}
   - component: {fileID: 114800310164848628}
   m_Layer: 0
   m_Name: AgentB
@@ -1177,15 +1179,15 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 0.5, y: 8, z: 11}
   m_Center: {x: 0, y: 0, z: 0}
---- !u!114 &114800310164848628
+--- !u!114 &114176423636690854
 MonoBehaviour:
   m_ObjectHideFlags: 1
   m_PrefabParentObject: {fileID: 0}
   m_PrefabInternal: {fileID: 100100000}
-  m_GameObject: {fileID: 1882383181950958}
+  m_GameObject: {fileID: 1170495812642400}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: e51a3fb0b3186433ea84fc1e0549cc91, type: 3}
+  m_Script: {fileID: 11500000, guid: 5d1c4e0b1822b495aa52bc52839ecb30, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
   m_BrainParameters:
@@ -1199,6 +1201,39 @@ MonoBehaviour:
   m_InferenceDevice: 0
   m_UseHeuristic: 0
   m_BehaviorName: Tennis
+--- !u!114 &114399072728845634
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1882383181950958}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 5d1c4e0b1822b495aa52bc52839ecb30, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  m_BrainParameters:
+    vectorObservationSize: 8
+    numStackedVectorObservations: 3
+    vectorActionSize: 02000000
+    cameraResolutions: []
+    vectorActionDescriptions: []
+    vectorActionSpaceType: 1
+  m_Model: {fileID: 11400000, guid: c85010de4f32e4c88bac16d9688aaadc, type: 3}
+  m_InferenceDevice: 0
+  m_UseHeuristic: 0
+  m_BehaviorName: Tennis
+--- !u!114 &114800310164848628
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1882383181950958}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e51a3fb0b3186433ea84fc1e0549cc91, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   agentParameters:
     agentCameras: []
     agentRenderTextures: []
@@ -1236,17 +1271,6 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e51a3fb0b3186433ea84fc1e0549cc91, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  m_BrainParameters:
-    vectorObservationSize: 8
-    numStackedVectorObservations: 3
-    vectorActionSize: 02000000
-    cameraResolutions: []
-    vectorActionDescriptions: []
-    vectorActionSpaceType: 1
-  m_Model: {fileID: 11400000, guid: c85010de4f32e4c88bac16d9688aaadc, type: 3}
-  m_InferenceDevice: 0
-  m_UseHeuristic: 0
-  m_BehaviorName: Tennis
   agentParameters:
     agentCameras: []
     agentRenderTextures: []
