--- UnitySDK/Assets/ML-Agents/Scripts/Policy/PolicyFactory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1e63954c62d854d8db4d94dc41269441
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
